Added support for worklets in closure
To be able to use other worklets in closures we need to have support for detecting and calling worklets that are passed in a closure to the wrapper functions.

This commit adds support for doing this by letting the wrapper system know about the worklet context so that it can create additional worklets when seeing them defined.